Refactored Cookie Consent System
Refactored the cookie consent system due the non flexibility for other cookies, so now it will be easier to either remove or add new cookies and to store them appropriately.

Added more translations, added a common translation for more "common" elements such as the title (more to come I promise).