Feature/notifications (#331)
* [frontend] Changed notification click event

Also added the language to the POST request when subscribing to notifications. Minor tweak to the update text.

* [backend] Added `google-cloud-tasks`

Initialized the client for it

* [backend] New Model for Discord messages

Storing their ID alongside what type of message it is, event -> event_id, news -> news_id

* [backend] Language query for notifications

* [backend] Added FKs to Notifications

This is more aligned towards notifications for the web we don't want to continue fetching non-existing notifications. Also added a hybrid property to get the end_date for events

* [backend] Added services

Added utility services for sending Pub/Sub messages, Discord webhook messages and scheduling with GCP Cloud Tasks.

* [backend] Updated notification service

Added a new function to send push notifications and adjusted the notification creation to now support the new FKs and to remove the unnecessary if statement

* [backend] Adjusted the item service

Removed duplicate name detection and we needed the language code later on past the function call so we had to copy the translation data without the language_code.

* [backend] Created a new tasks route

Hopefully this works as intended.

* [backend] Adjusted scheduler route

It now uses the new send_notification service function. Optimized the event querying stuff, it will now query 23:59 before an event and get the correct committee you have subscribed to.

* Update message_routes.py

* [backend] Added the hooks for the notifications

It will schedule the news message to Discord so that it hopefully can catch the open graph visuals before sending the link, and it will defer the task of adding the notification to the Pub/Sub.

For events it is a lot simpler, it will simply just send a message to Discord and add a notification entry via Pub/Sub

* [backend] Registering the new route

* Update news_routes.py

* [backend] Last touchups and bugfixes