Task/security improvements (#332)
* [frontend] Rewrite the API URL

Instead of using the constant variable, making it easier to maintain and more obscure.

* [frontend] Add Discord redirect URL

* [frontend] Redesigned the middleware

Separated the concern and increased the modularity.

* [project] fixed bugs

Just strange bugs with trailing slashes that occurred due to the rewrite in the next config.

* [project] Adjusted the auth provider

It is now more tolerant and won't 401 when not authenticated.

* [backend] Another bugfix

It supplied UUIDs instead of strings causing it to not serialize.

* [frontend] CalendarProvider bugfix

Will now use the rewrite instead.