More refinement for RGBank (#340)
* Update RGBank (#335)

* [frontend] Fixing Bugs

Finally fixing the carousel bug and updated the versions, also address some api issues

* Added a way to track sent notifications

* [backend] Fixed the scheduler

Improved the discord notifications

* [frontend] From user feedback

Changed the search button to a quick access button, since the search was problematic on phones for users. This is a placeholder for now (might change), until I can design something better potentially.

* Updated the manifest (#334)

To build a better PWA experience

* Moving stuff around

* Added authentication

From main-app, will be in the shared folder later.

* Added a main provider wrapper

Moved the auth provider in to that instead of the client wrapper, since the header needs it. Added the beginnings of the login section of the header.

* Update the lock file

* New Components

* Account page

A basic design, will need to add more stuff

* Specific expense details page

* Added the wrapper for next-view-transitions

* Additional conditions

Only one part/committee can be allowed to be on an invoice, also added a description to the expense form.

* Added basic auth protection

* Added middleware and authentication

* Start of backend models

* Added base admin pages

* [backend] Update the models

Improved the enum, so that it now correctly assigns to the schema and added a hybrid property to invoices and expenses to retrieve the committee if it is present.

* [backend] Initial routes for the new app

Also added a basic service file to assist the new routes

* [backend] Fixed the JSONB

* [frontend] Removed the type input from categories

* [frontend] Updated forms and finalization

Added eventual language support

* [backend] Added different bucket support

* [backend] Improve the clarity of routes

* [project] Semi-working invoice detail view

Get all dynamic data for each invoice, still need to add comments and more in-depth user details.

* [project] Messages work

* You can now comment on expenses and invoices that have been uploaded.
* Improved the account history page
* Other UI improvements

* [backend] New models

Mainly for statistic counting

* [backend] Statistics routes

It will update the relevant statistics when the status of the (`invoice` | `expense`) is updated to BOOKED.

* [frontend] Updated Next.js

* [frontend] API hooks

Also solved a 204 bug

* [project] Updated the global next.js version

* [frontend] Fully functioning statistics page

* [project] Basic auth setup

* [backend] Forgot to add the decorator

* [backend] Reduce the docker image size

Using a minifier and removing some excess files that might be added in the image

* [project] Simple auth check in middleware

Might expand later in to more in depth checks.

* [backend] First implementation of mails!

Can be used elsewhere, albeit with limitations (100 mails/day)*

* for free at least.

* [backend] More permissions and separated logic

* [frontend] Added better logic

You will now return to the previous page if you visit a protected route, useful for email links that might link to when the recipient is not logged in. Some basic auth checks for certain actions.

* [frontend] Bug fix and separated check

Hopefully this fixes a bug where the expense is unregistered when the user is away for long enough. Also separated the check for whether you can access the admin page.

* [project] Domains works now

You can create update and rename domains and use them when creating an expense/invoice

* [frontend] Responsive & improve language

* [frontend] Better header & dark mode

* [backend] Reception mode safe

* [backend] New routes and modified `to_dict`

Reduced the payload when requesting all data.

* [frontend] Added Metadata

* [frontend] Reflected the model/responses

* [frontend] Changed the expense badges

* [frontend] Improved login/logout flow

* [frontend] New Admin page

Added a way to track ongoing expenses

* [frontend] Improved error handling

* [backend] Fixed DB error

And added title column

* [frontend] New attributes and bug fixes

Fixed login issues, and added the title attribute from invoices and expenses to the UI, also moved the files around

* [project] Removed statistics and others

Added the title to each form, and updated the form to fix the description being updated on expenses. Removed the individual statistics due to potential concern of "competing". Also fixed some UI issues.

* [project] Added Peacock colors for VSCode

* [backend] Adjusted the imports

* [project] Admins can change categories

* [frontend] Tests!

Added basic tests, and added the title attribute to the cards

* [frontend] Translations

* [frontend] Production ready

Fixed ESLint issues and minor bug fixes

* [backend] Minor fixes

Also added a healthcheck for Docker, and coupled the mail, will change to the real mail after

* [frontend] Moved unit tests

* [frontend] Fixing middleware redirects

* [frontend] E2E Tests with Playwright

And much more accessibility attributes

* [backend] Rich stacktraces

And minor tweaks

* [project] 1️⃣ Creating shared packages

* [project] 2️⃣ Creating shared packages

Moving API and basic UI util to shared package.
Updated the UI components to the latest versions, I tried moving the components to a shared package but it didn't work well.

* [frontend] Dropping shadcn/ui forms

They are incredibly buggy in combination with hookresolvers, causing insane typescript issues. Upgraded to zod@mini v4

* [frontend] Zod bug fix

* [backend] Fixed mailing

* [project] Added more details to items

Added more details when an item has been booked such as the person who booked the item and when.

* [project] Fixed Zod errors by upgrading

* [frontend] Updated the auth

* [frontend] Production ready

* Added a requirement to upload a expense/invoice to require you to select which file belongs to which category
* Updated tests to reflect this new requirement
* Minor UI updates to files in the detailed view

* [backend] Production ready

* Adjusted the reception mode dict for students
* Adjusted the creation of expenses/invoices to reflect new requirements in the frontend
* Adjusted the auth service to not include past positions
* Adjusted the mail service to take the not public student dict (for reception mode)

* [frontend] Last issues fixed

Added a way for admins to change the category files!

* [project] Update turbo to version 2.5.3 in root package.json

fix: update next dependency to version 15.3.3 in middleware package.json

* [project] Refactor middleware and update dependencies; enhance response handling and schema validation